Added confirm for deleting room

+++ src/app/admin/rooms/room-detail/room-detail.component.ts
@@ -29,16 +29,19 @@ export class RoomDetailComponent implements OnInit {
   }
 
   deleteRoom() {
-    this.message = 'deleting...';
-    this.dataService.deleteRoom(this.room.id).subscribe(
-      next => {
-        this.dataChangedEvent.emit();
-        this.router.navigate(['admin', 'rooms']);
-      },
-      error => {
-        this.message = 'Sorry - this room cannot be deleted at this time';
-      }
-    );
+    const result = confirm('Are you sure you wish to delete this room?');
+    if (result) {
+      this.message = 'deleting...';
+      this.dataService.deleteRoom(this.room.id).subscribe(
+        next => {
+          this.dataChangedEvent.emit();
+          this.router.navigate(['admin', 'rooms']);
+        },
+        error => {
+          this.message = 'Sorry - this room cannot be deleted at this time';
+        }
+      );
+    }
   }
 
 }
